moved signal handler and some fixes

+++ src/process.cpp
@@ -14,7 +14,17 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 
-#include "process.hpp"
+#include <process.hpp>
+
+std::atomic<bool> Process::isActive_;
+
+bool Process::isActive() {
+	return isActive_;
+}
+
+void Process::setActive(bool isActive) {
+	isActive_ = isActive;
+}
 
 int Process::createPid(std::string pidPath) {
 	pidPath_ = pidPath;
@@ -37,10 +47,15 @@ int Process::createPid(std::string pidPath) {
 }
 
 void Process::destroyPid() {
-	flock(pidFd_, LOCK_UN);
-	close(pidFd_);
-	unlink(pidPath_.c_str());
-	hasPid_ = false;
+	if (pidFd_) {
+		flock(pidFd_, LOCK_UN);
+		close(pidFd_);
+	}
+
+	if (hasPid_) {
+		unlink(pidPath_.c_str());
+		hasPid_ = false;
+	}
 }
 
 int Process::createFifo(std::string fifoPath) {
@@ -51,30 +66,63 @@ int Process::createFifo(std::string fifoPath) {
 		return 1;
 	}
 
-	std::cerr << fifoPath << " has been created." << std::endl << "Waiting for user to open it." << std::endl;
-	fifoFd_ = open(fifoPath_.c_str(), O_WRONLY);
+	// TODO unblock, when signal has been received
 	hasFifo_ = true;
+	std::cerr << fifoPath << " has been created. Waiting for user to open it." << std::endl;
+	fifoFd_ = open(fifoPath_.c_str(), O_WRONLY);
 
 	return 0;
 }
 
 void Process::destroyFifo() {
-	unlink(fifoPath_.c_str());
-	close(fifoFd_);
-	hasFifo_ = false;
+	if (fifoFd_) {
+		close(fifoFd_);
+	}
+
+	if (hasFifo_) {
+		unlink(fifoPath_.c_str());
+		hasFifo_ = false;
+	}
+}
+
+int Process::getFifoFd() {
+	return fifoFd_;
+}
+
+// TODO stop GCHD aswell, when signal has been received
+void Process::sigHandler_(int sig) {
+	std::cerr << std::endl << "Stop signal received." << std::endl;
+
+	switch(sig) {
+		case SIGINT:
+			isActive_ = false;
+			break;
+		case SIGTERM:
+			isActive_ = false;
+			break;
+	}
 }
 
 Process::Process() {
+	isActive_ = false;
 	hasPid_ = false;
 	hasFifo_ = false;
+	pidFd_ = 0;
+	fifoFd_ = 0;
+
+	// signal handling
+	struct sigaction action;
+	action.sa_handler = sigHandler_;
+	sigaction(SIGINT, &action, nullptr);
+	sigaction(SIGTERM, &action, nullptr);
+
+	// ignore SIGPIPE, else program terminates on unsuccessful write()
+	struct sigaction ignore;
+	ignore.sa_handler = SIG_IGN;
+	sigaction(SIGPIPE, &ignore, nullptr);
 }
 
 Process::~Process() {
-	if (hasFifo_) {
-		destroyFifo();
-	}
-
-	if (hasPid_) {
-		destroyPid();
-	}
+	destroyFifo();
+	destroyPid();
 }
